Add frontend router; Add login and forget password page
