+++ mit_frontend/src/routers/ResetPassword.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import MuiThemeProvider from 'material-ui/styles/MuiThemeProvider';
+import {Card, CardTitle} from 'material-ui/Card';
+import TextField from 'material-ui/TextField';
+import RaisedButton from 'material-ui/RaisedButton';
+import '../App.css';
+
+class ResetPassword extends Component {
+
+}
+
+export default ResetPassword;
