+++ mit_frontend/src/routers/styles.js
@@ -1,0 +1,21 @@
+const styles = {
+  textField: {
+    text: {
+      color: '#3F51B5'
+    },
+    underline: {
+      borderColor: '#3F51B5'
+    },
+    error: {
+      color: '#bf2a5c'
+    }
+  },
+  button: {
+    color: '#4285F4',
+    label: {
+      color: '#ffffff'
+    }
+  }
+}
+
+export default styles;
